Backend/pyproject.toml
[project]
name = "resume-analyzer-backend"
version = "0.1.0"
description = "Resume Analyzer Backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
]

[tool.uv]
dev-dependencies = []
